refactor: HighlightText, Comment 컴포넌트 분리
